requirements.txt
google-genai
semanticscholar
rich
python-dotenv
requests